// files-manager-web/src/api/generated.ts
export enum ConflictPolicyDto {
    Rename = "Rename",
    Overwrite = "Overwrite",
    Skip = "Skip",
    Error = "Error",
}

export enum LogLevel {
    Trace = "Trace",
    Debug = "Debug",
    Information = "Information",
    Warning = "Warning",
    Error = "Error",
    Critical = "Critical",
    None = "None",
}
// ...
export interface ISettingDto {
    watchFolder?: string | undefined;
    logLevel?: LogLevel;
    conflictPolicy?: ConflictPolicyDto;
}